backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
mysqlclient
django-environ
